Constant streaming, Log file structure, Stability
- Improved the format of the log file. Removed calibration status text,
the test wont start unless the sensor is calibrated anyways.
- If you press the escape button at any point apart from the end screen,
the program closes and the current log file is deleted. Provided
addition junk file creation.
- Changed logging from Python from when a bool is true, to constant.
This means the we will always be able to get the starting head position
when the task starts as it removes any data lag.
- Forced game into only allowing one instance at a time, providing
greater stability if there is a crash.